% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catbot.r
\name{add_cat_sep}
\alias{add_cat_sep}
\title{Add formatted separators to R output with cats!}
\usage{
add_cat_sep(opt = 1L, width = 58L)
}
\arguments{
\item{opt}{Separator style, supplied as a number or name.}

\item{width}{Total width of the separator, including brackets where used.}
}
\value{
The printed separator invisibly.
}
\description{
Add formatted separators to R output with cats!
}
\examples{
add_cat_sep(
  opt = "cat2"
)

add_cat_sep(
  opt = "cat1",
  width = 58L
)

}
