% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_custom_formatting.r
\name{add_clock}
\alias{add_clock}
\title{Add the current date and time in a box}
\usage{
add_clock(
  time = Sys.time(),
  date_format = "\%Y-\%m-\%d",
  time_format = "\%H:\%M:\%S",
  width = 12L,
  style = c("unicode", "ascii")
)
}
\arguments{
\item{time}{A date-time object. Defaults to the current time.}

\item{date_format}{Format passed to \code{\link[base:format]{base::format()}} for the date.}

\item{time_format}{Format passed to \code{\link[base:format]{base::format()}} for the time.}

\item{width}{Width of the box interior.}

\item{style}{Box style passed to \code{\link[=add_box]{add_box()}}.}
}
\value{
The printed box invisibly as a character vector.
}
\description{
Add the current date and time in a box
}
\examples{
add_clock()

add_clock(
  date_format = "\%d/\%m/\%Y"
)
}
