% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_custom_formatting.r
\name{add_h}
\alias{add_h}
\title{Add custom headers to R scripts}
\usage{
add_h(opt = 1L, text)
}
\arguments{
\item{opt}{Heading style, supplied as a number or name.}

\item{text}{text string for the header}
}
\value{
the printed header
}
\description{
Add custom headers to R scripts
}
\examples{
add_h(
 opt = 1,
 text = "OBJECTIVE COMPLETE: ANALYSIS COMPLETE"
)

add_h(
 opt = "warning",
 text = "OBJECTIVE COMPLETE: ANALYSIS COMPLETE"
)

}
