% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_custom_formatting.r
\name{add_sep}
\alias{add_sep}
\title{Add formatted separators to R output}
\usage{
add_sep(opt = 1L, width = 58L)
}
\arguments{
\item{opt}{Separator style, supplied as a number or name.}

\item{width}{Total width of the separator, including brackets where used.}
}
\value{
The printed separator invisibly.
}
\description{
Add formatted separators to R output
}
\examples{
add_sep(
  opt = "block"
)

add_sep(
  opt = "dash",
  width = 58L
)

add_sep(
  opt = 3L,
  width = 58L
)

}
