% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catbot.r
\name{catbot_say}
\alias{catbot_say}
\title{Print a small catbot message}
\usage{
catbot_say(
  text,
  style = c("robocat", "catfloat", "catbox"),
  width = 24L,
  align = c("center", "left", "right")
)
}
\arguments{
\item{text}{Character vector containing one or more message lines.}

\item{style}{Catbot style: \code{"robocat"}, \code{"catfloat"}, \code{"catbox"}}

\item{width}{Width of the text area when \code{style = "catbox"}.}

\item{align}{Text alignment when \code{style = "catbox"}:
\code{"center"}, \code{"left"} or \code{"right"}.}
}
\value{
Invisibly returns the complete printed output as a character vector.
}
\description{
Prints a catbot-style message using a robot cat, floating cat, or catbox
}
\examples{
catbot_say(
  text = "ANALYSIS COMPLETE"
)

catbot_say(
  text = "DATA CHECK COMPLETE",
  style = "catbox"
)


}
