% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_session.r
\name{.git_run}
\alias{.git_run}
\title{Run a Git command in a project}
\usage{
.git_run(project, args)
}
\arguments{
\item{project}{Project directory.}

\item{args}{Character vector of Git arguments.}
}
\value{
A list containing command output and exit status.
}
\description{
Run a Git command in a project
}
\keyword{internal}
